Adding better error handling to contact endpoints

+++ src/contacts/data/repositories/contactRepository.js
@@ -1,4 +1,5 @@
 import { getFirestore } from 'firebase-admin/firestore';
+import { contactErrors } from '../../domain/errors/contactErrors.js';
 
 export const contactExists = async (userId, contactId) => {
     const db = getFirestore();
@@ -25,14 +26,10 @@ export const saveContact = async (userId, contactId, contact) => {
 
 export const updateContact = async (userId, contactId, phone, address) => {
     const db = getFirestore();
-
-    console.log(userId);
-    console.log(contactId);
     const updatedContact = {
         phone: phone,
         address: address
     }
-
     const docRef = db
         .collection('users')
         .doc(userId)
@@ -43,12 +40,20 @@ export const updateContact = async (userId, contactId, phone, address) => {
 };
 
 export const deleteContact = async (userId, contactId) => {
-    const db = getFirestore();
-    const docRef = db
-        .collection('users')
-        .doc(userId)
-        .collection('contacts')
-        .doc(contactId);
+    const docExists = await contactExists(userId, contactId);
 
-    await docRef.delete();
+    if (docExists) {
+        const db = getFirestore();
+        const docRef = db
+            .collection('users')
+            .doc(userId)
+            .collection('contacts')
+            .doc(contactId);
+
+        await docRef.delete();
+    } else {
+        const error = new Error('Contact missing');
+        error.code = contactErrors.CONTACT_NOT_FOUND;
+        throw error;
+    }
 };
